requirements.txt
# Minimal runtime dependencies for Razorpay payment QR flow
Pillow>=9.0.0
requests>=2.28.0
razorpay>=2.0.0
